+++ Arduino/main/HelperFunctions.h
@@ -1,6 +1,6 @@
 #ifndef HelperFunctions_h
 #define HelperFunctions_h
 
-float lerp(float a, float b, float x);
+int lerp(int a, int b, int x);
 
 #endif
